refactor(pubmed_ingest): moved files to simplify dir structure
